Add author and category tag to cards
To show more information to the user, we can add the author which might
call their attention to interact with the card.
Moreover we can add a tag that specifies at a high level what the card
is about to let them easily grasp if it will be of their interest or
not.

This commit adds the styling to show the author and a mocked tag with
randomized categories.