Improve overall page
This commit adds multiple improvements to our existing implementation.
It adds a loading state for when we are fetching cards.
It adds a random tag generation to tag the content of the cards.
It adds an empty state for when there are no related cards.
It also adds some styles that we would improve later on.
Finally, it modifies the existing typography and implements the
initial setup to render them.